+++ config.go
@@ -1,16 +1,14 @@
 package main
 
 import (
+	"fmt"
 	"io/ioutil"
-	"regexp"
 
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v3"
 )
 
-type regex struct {
-	regexp.Regexp
-}
+// STRUCTS
 
 type profile struct {
 	Source  string              `yaml:"source"`
@@ -18,22 +16,36 @@ type profile struct {
 	Modules []map[string]string `yaml:"modules,omitempty"`
 }
 
+type mailConfig struct {
+	SMTPServer string `yaml:"smtp_server"`
+	SMTPPort   int    `yaml:"smtp_port"`
+	Sender     string `yaml:"sender"`
+	SMTPUser   string `yaml:"smtp_user,omitempty"`
+	SMTPPass   string `yaml:"smtp_pass,omitempty"`
+}
+
 type serverConfig struct {
-	Addr     string    `yaml:"addr"`
-	LogLevel log.Level `yaml:"loglevel"`
+	Addr        string     `yaml:"addr"`
+	URL         string     `yaml:"url"`
+	LogLevel    log.Level  `yaml:"loglevel"`
+	StoragePath string     `yaml:"storagepath"`
+	Mail        mailConfig `yaml:"mail,omitempty"`
+}
+
+type notifier struct {
+	Source     string   `yaml:"source"`
+	Interval   string   `yaml:"interval"`
+	Recipients []string `yaml:"recipients"`
 }
 
 // Config represents configuration for the application
 type Config struct {
-	Profiles map[string]profile `yaml:"profiles"`
-	Server   serverConfig       `yaml:"server"`
+	Server    serverConfig        `yaml:"server"`
+	Profiles  map[string]profile  `yaml:"profiles,omitempty"`
+	Notifiers map[string]notifier `yaml:"notifiers,omitempty"`
 }
 
-func (r *regex) UnmarshalText(text []byte) error {
-	tmpRe, err := regexp.Compile("(?i)" + string(text))
-	r.Regexp = *tmpRe
-	return err
-}
+// CONFIG MANAGEMENT FUNCTIONS
 
 // ParseConfig reads config from path and returns a Config struct
 func ParseConfig(path string) (Config, error) {
@@ -41,12 +53,14 @@ func ParseConfig(path string) (Config, error) {
 
 	yamlFile, err := ioutil.ReadFile(path)
 	if err != nil {
-		log.Printf("yamlFile.Get err   #%v ", err)
+		log.Fatalf("Error Reading Config: %v ", err)
+		return tmpConfig, err
 	}
 
 	err = yaml.Unmarshal(yamlFile, &tmpConfig)
 	if err != nil {
-		log.Fatalf("Unmarshal: %v", err)
+		log.Fatalf("Error Unmarshalling Config: %v", err)
+		return tmpConfig, err
 	}
 
 	// defaults
@@ -56,6 +70,68 @@ func ParseConfig(path string) (Config, error) {
 	if tmpConfig.Server.LogLevel == 0 {
 		tmpConfig.Server.LogLevel = log.InfoLevel
 	}
+	if tmpConfig.Server.StoragePath == "" {
+		tmpConfig.Server.StoragePath = "/etc/ical-relay/"
+	}
 
 	return tmpConfig, nil
+}
+
+func reloadConfig() error {
+	// load config
+	var err error
+	conf, err = ParseConfig(configPath)
+	if err != nil {
+		return err
+	} else {
+		log.Info("Config reloaded")
+		return nil
+	}
+}
+
+func (c Config) saveConfig(path string) error {
+	d, err := yaml.Marshal(&c)
+	if err != nil {
+		return err
+	}
+	return ioutil.WriteFile(path, d, 0644)
+}
+
+// CONFIG EDITING FUNCTIONS
+
+func (c Config) getPublicCalendars() []string {
+	var cal []string
+	for p := range c.Profiles {
+		if c.Profiles[p].Public {
+			cal = append(cal, p)
+		}
+	}
+	return cal
+}
+
+func (c Config) addNotifyRecipient(notifier string, recipient string) error {
+	if _, ok := c.Notifiers[notifier]; ok {
+		n := c.Notifiers[notifier]
+		n.Recipients = append(n.Recipients, recipient)
+		c.Notifiers[notifier] = n
+		return c.saveConfig(configPath)
+	} else {
+		return fmt.Errorf("notifier does not exist")
+	}
+}
+
+func (c Config) removeNotifyRecipient(notifier string, recipient string) error {
+	if _, ok := c.Notifiers[notifier]; ok {
+		n := c.Notifiers[notifier]
+		for i, r := range n.Recipients {
+			if r == recipient {
+				n.Recipients = append(n.Recipients[:i], n.Recipients[i+1:]...)
+				c.Notifiers[notifier] = n
+				return c.saveConfig(configPath)
+			}
+		}
+		return fmt.Errorf("recipient not found")
+	} else {
+		return fmt.Errorf("notifier does not exist")
+	}
 }