+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"net/http"
 	"os"
 
@@ -8,36 +9,57 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-var version = "1.1.6"
+var version = "1.2.0"
 
-var configPath = "config.yml"
+var configPath string
 var conf Config
 
 var router *mux.Router
 
 func main() {
-	if len(os.Args) >= 2 {
-		configPath = os.Args[1]
-	}
-
 	log.Info("Welcome to ical-relay, version " + version)
+
+	var notifier string
+	flag.StringVar(&notifier, "notifier", "", "Run notifier with given ID")
+	flag.StringVar(&configPath, "config", "config.yml", "Path to config file")
+	flag.Parse()
 
 	// load config
 	var err error
 	conf, err = ParseConfig(configPath)
 	if err != nil {
-		log.Fatal(err)
+		os.Exit(1)
 	}
 
 	log.SetLevel(conf.Server.LogLevel)
 	log.Debug("Debug log is enabled") // only shows if Debug is actually enabled
 
-	// setup router
+	// run notifier if specified
+	if notifier != "" {
+		log.Debug("Notifier mode called. Running: " + notifier)
+		err := RunNotifier(notifier)
+		if err != nil {
+			os.Exit(1)
+		} else {
+			os.Exit(0)
+		}
+	} else {
+		log.Debug("Server mode.")
+	}
+
+	// setup routes
 	router = mux.NewRouter()
 	router.HandleFunc("/", indexHandler)
 	router.HandleFunc("/profiles/{profile}", profileHandler).Name("profile")
+	router.HandleFunc("/api/calendars", calendarlistApiHandler)
+	router.HandleFunc("/api/reloadconfig", reloadConfigApiHandler)
+	router.HandleFunc("/api/notifier/{notifier}/addrecipient", addNotifyRecipientApiHandler).Name("notifier")
+	router.HandleFunc("/api/notifier/{notifier}/removerecipient", removeNotifyRecipientApiHandler).Name("notifier")
 
-	// listen and serve
+	// start notifiers
+	NotifierStartup()
+
+	// start server
 	address := conf.Server.Addr
 	log.Fatal(http.ListenAndServe(address, router))
 }